Create SGPA Calculation Program By Arrays and Functions with Validations
Problem Statement:
Write a program to calculate the semester grade point average (SGPA) of the student. Declare a const integer size
and initialize with the value 8. Now declare the following variables and arrays of the given size inside the main
function.
• Courses (An array of string data type to store the name of course like PF, PF_Lab etc.)
• CrHrs (An array of integer data type to store the credit hours of the courses.)
• Marks (An integer array to store marks in the range 0 - 100 in each course)
• Grades (An array of string data type to store the grade of each course
• Scores (An array of float data type to store the score out of 4 in each subject. You don’t need to
prompt the user to enter this score rather use the mapping provided at the end of document).
• sgpa (a float variable to store the semester gpa) You need to calculate the sgpa by using the data
of given arrays.
• crErnd (an int variable) to store the credit earned by the student. Credits earned means number of
credit hours passed by the student. If a student failed to obtain 50 marks in any particular course,
then the credit hours of that course will not be considered as earned credit hours.
Create a function getData and pass the arrays (Courses, CrHrs, Marks) and the int variable size as parameters to
from main function. This function should prompt the user to enter the data. Store the data on relevant indices i.e.,
if you are storing “PF” on index 0 of courses then store the credit hours and marks in “PF” on the index location
‘0’ of CrHrs and marks array.
Create a function AssignGrades and pass the arrays (Marks, Grades, and Scores) and the int variable size as
parameters from main function. This function should assign the grades and the scores as per the mapping
provided at the end of document.

Create a function getSGPA and pass the arrays (Scores and CrHrs), and the int variables size, sgpa and crErnd(by-
reference) from main function. This function should calculate and store credit hours earned by student in a

variable crErnd and semester grade point average in sgpa(parameter received by reference). To calculate the
SGPA you can use the following approach:
SGPA = Summation of the product of corresponding indices of the given arrays i.e., (Scores and CrHrs) divided

by totalCrHrs.

Create a function printData and pass the arrays (courses, marks, grades), and the variables size, crErnd, and sgpa
of int and float data type respectively from main function. This function should print the data on console.
Note: Provide generic approach to solve the given problem and use loops wherever needed. No reward for
hardcoded program. Your program repeatedly prompts the user to enter valid data in case of invalid input.