Create Find Equilibrium Index in a 1D Static Int Array
Write a program that declares a 1D static integer array and get the input from user. Your task is to
determine the equilibrium index of the array. As discussed in the class, sum of values on the left of
equilibrium index should be equal to the sum of values on the right side. Input values should be in
the range (-5000 to 5000).