Create Sudoku Game in C++
Sudoku is one of the most popular classic games. The playing grid of sudoku consists of the
9*9 squares space. The game initially has 9 squares combining rows and columns. Out of
these 9 squares, every square is a separate grid of 3*3 square spaces. Every row and
column consist of 9 squares. It can be either one player or a two-player game where the
competitors try to fill the boxes faster than their opponents. The game’s primary purpose is
to fill boxes in every square with numbers ranging from 1 to 9. The exciting thing about the
Sudoku game is that no column, row, or square allows repeating a number.
Moreover, the game initially fills some of the squares with random numbers from 1 to 9. The
rest of the boxes or squares are left empty for the player to fill. The number of already filled
boxes determines the complexity of the game. Following is a picture of a sample Sudoku
game for better visualization.

Requirements and Conditions of the Game
In this Sudoku game, the user wants to develop a game where the player fills empty boxes
with remaining numbers according to the game’s conditions. Following are the game
requirements and resultant conditions for this Sudoku game:
● The game menu ask user for input to either load game or start a new game
● The game board should be of 9*9 squares space.
● The game should display the grid with some of the boxes already filled with random
numbers ranging from 1 to 9.
● The game should allow the player to fill in empty boxes.
● The game should not allow the player to repeat any number in any row, column, or
square.
● If the player fills all the boxes with correct numbers and without any repetition, he
wins the game.
Load/Saving
You will use the simple sequential text file to store the moves, and at one time, the user
should be able to store a max of 3 different games if they want, and should

be able to load any one of these games and continue playing. You should maintain all the
files needed for this purpose internally, i.e., the user doesn’t need to know what these files
are and where they’re stored.
File format will be sequence of 81 numbers having random numeric and so space it would
be #.
Example input would be: 3 # # 8 # 1 # # 2 first row of the grid.