Create Sand Hour Glass in C++
Problem Statement:
Write a C++ program to simulate an hourglass. Get the following input from the user: Write a program to simulate hourglass. Get the following input from user. 
Time: Total time (in seconds) that can be measured/simulated using this hourglass. (Remember: Time can’t be 0 or negative) 
C: The character to represent sand in the hourglass. 
The simulation works in this way: after one second, some sand from the upper portion of the  clock drops into the lower portion of the clock. For your program, you can implement it in this  way: after every second, remove one row from the upper portion of the hourglass and add a  new row in the lower portion of the hourglass. The simulation stops when no sand remains in  the upper portion of the hourglass. You can use the Sleep function to stop execution of the  program for 1 second.