Create Filter Traversal on Image Array in C++
Write a C++ Program, where we Create a 2D Static Array, and call it Image. Create Another 2D static Array smaller than image and call it Filter. Then Traverse the Filter through 2D array so that it does not go outside the Boundary of Image, and Calculate the results to be stored in a resulting array. Results will be calculated by Multiplying the corresponding values of filter  and Image. Check Code for further Understanding...